Bug 1808605 - part 12: Rename `build` tranforms into `build_components`
