+++ taskcluster/ci/build-components/kind.yml
@@ -5,7 +5,7 @@
 loader: android_taskgraph.loader.build_config:components_loader
 
 transforms:
-    - android_taskgraph.transforms.build:transforms
+    - android_taskgraph.transforms.build_components:transforms
     - android_taskgraph.transforms.gradle_optimization:transforms
     - taskgraph.transforms.job:transforms
     - taskgraph.transforms.task:transforms
