+++ taskcluster/ci/build-samples-browser/kind.yml
@@ -5,7 +5,7 @@
 loader: taskgraph.loader.transform:loader
 
 transforms:
-    - android_taskgraph.transforms.build:transforms
+    - android_taskgraph.transforms.build_components:transforms
     - android_taskgraph.transforms.gradle_optimization:transforms
     - taskgraph.transforms.job:transforms
     - taskgraph.transforms.task:transforms
